Rename locationUtils to location
